Clean up gallery code
 - Make GalleryRowContainer the Gallery component
 - Move Gallery controls to GalleryModal
 - Remove types for unused react-photo-gallery
 - Move all gallery dimensions to enum
