+++ src/styles/gallery.less
@@ -1,14 +1,11 @@
 .gallery {
-}
-
-.gallery-row-container {
   display: flex;
   flex-direction: column;
 
   .gallery-row {
     display: flex;
     flex-direction: row;
-    justify-content: space-around;
+    justify-content: space-between;
     align-items: center;
 
     .thumbnail {
